Update : Display variations with missing media
(Cherry picked from main, books.edit view not merged)

We need to display variations without media to warn user that the varia-
tion won't be displayed on the front-end. That  occurs when a media is
unlinked from its variation in the media library.

We had to drop the has() method in the relation definition. That was a
mess. It was adding a where clause in the query builder instance, but
we couldn't edit it when needed, even with getQuery(), because the
'wheres' property is protected.

When used as an eloquent query builder directly in the list() of the
BooksController, books without media were filtered out only if all the
books were missing media. If one book had media, all books was returned
and Glide.js was hanging obviously.

Now in list(), we query for all bookInfos with its books and all their
media, then filter out books which has no media.

In edit(), we add withTrashed and withMedia to the relationship query
then execute it, and then apply a partition on the resulting collection